Fix not redirecting issue
