+++ front-end/peer-prep/src/components/UserPage/index.tsx
@@ -22,7 +22,9 @@ const UserPage = () => {
 
     // Gets user details from firebase.
     const user = auth.currentUser;
-    const authEmail = user?.providerData[0].email ?? ""
+    //const authEmail = user?.providerData[0].email ?? ""
+    console.log(user)
+    const authEmail = user?.email ?? ""
     const authUsername = user?.displayName ?? ""
     const authUid = user?.uid
     
@@ -50,7 +52,6 @@ const UserPage = () => {
                 console.log(error)
                 dispatch(UserSlice.updateCurrentEmail(authEmail))
                 dispatch(UserSlice.updateCurrentUsername(authUsername))
-                dispatch(UserSlice.setIsFirstTimeLogin(true))
         })
     },[])
