+++ front-end/peer-prep/src/index.tsx
@@ -5,7 +5,7 @@ import { auth } from "./components/Auth/Firebase";
 import { useAuthState } from "react-firebase-hooks/auth";
 
 // import Redux components here
-import { Provider, useSelector } from "react-redux";
+import { Provider, useDispatch, useSelector } from "react-redux";
 import store from "./components/redux/store/store"
 import * as UserSlice from "./components/redux/reducers/User/UserSlice"
 
@@ -29,6 +29,8 @@ import SignInPage from "./components/Auth/SignInPage";
 import Navbar from "./components/Navbar";
 import GoodbyePage from "./components/Auth/GoodbyePage";
 
+import axios from "axios";
+
 const ProtectedRoute = () => {
 	const [user, loading, error] = useAuthState(auth);
 	debugger;
@@ -49,10 +51,21 @@ const ProtectedRoute = () => {
 };
 
 const RedirectUserRoute = () => {
+    const user = auth.currentUser;
+    const uid = user?.uid
     const isNewUser = useSelector(UserSlice.selectIsFirstTimeLogin);
+    const dispatch = useDispatch();
     debugger;
+    axios({
+        method: 'get',
+        url: `http://api.peerprepgroup51sem1y2023.xyz/users/${uid}`
+    }).catch((error) => {
+        console.log(error);
+        console.log("test");
+        dispatch(UserSlice.setIsFirstTimeLogin(true));
+    })
     if (isNewUser) {
-        debugger;
+        debugger;	
         // If user has not entered user details, navigate to User page
         return <Navigate to="/user" replace />;
     } else {
